Dommages aux ennemis j'arrive pas à savoir si ça marche

+++ Assets/scripts/Data/Monsters.cs
@@ -9,7 +9,7 @@ public class Monsters : ScriptableObject
     #region Attributes
 
     public string name;
-    public float lifePoints;
+    public int lifePoints;
     public float speed;
     public int strength;
     public float magicResistance;
